adding PrepNuQCWorkflow (#132)
* inserting data to qiita for testing

* testing creation of study in qiit

* add sbatch

* shopt

* echo .profile

* slurm

* partition

* scontrol create partition

* mv scontrol

* partitionname

* sudo scontrol

* add some prints

* /usr/bin/sbatch

* sudo

* env

* .local/bin/sbatch

* ls

* sbatch in conda

* squeue

* improve error display and running tests

* sbatch

* GITHUB_PATH

* adding files to tests/bin

* test_metagenomic_workflow_creation

* adding _inject_data

* fixing some tests

* fixing other tests

* copyfile -> touch()

* more copyfile -> touch()

* fix test_tellseq_workflow_creation

* addressing @AmandaBirmingham comments

* flake8

* adding main changes

* adding PrepNuQCWorkflow

* small changes before deploy

* metagenomics_pooling_notebook -> kl-metapool

* minor changes from current version in qiita

* fix TestMultiQCJob.test_generate_job_scripts

* hasattr fsr

* use run_prefix as sample_name

* fix test_PrepNuQCWorkflow

* addressing @AmandaBirmingham comments

* fix _generate_dummy_sample_sheet

* add missing ,

* add other missing ,

+++ src/qp_klp/Workflows.py
@@ -1,5 +1,5 @@
 from os.path import join, exists, split
-from os import walk, makedirs, listdir
+from os import walk, makedirs, listdir, environ
 import pandas as pd
 from json import dumps
 from subprocess import Popen, PIPE
@@ -759,6 +759,13 @@ class Workflow():
 
     def determine_steps_to_skip(self):
         out_dir = self.pipeline.output_path
+
+        # check if the test flag is on!
+        test = False
+        if 'PrepNuQCJob_TEST' in environ and \
+                environ['PrepNuQCJob_TEST'] == 'true':
+            test = True
+
         # Although amplicon runs don't perform host-filtering,
         # the output from ConvertJob is still copied and organized into
         # a form suitable for FastQCJob to process. Hence the presence or
@@ -776,6 +783,7 @@ class Workflow():
                             break
                     self.skip_steps.append(directory)
                 else:
-                    # work stopped before this job could be completed.
-                    rmtree(join(out_dir, directory))
+                    if not test:
+                        # work stopped before this job could be completed.
+                        rmtree(join(out_dir, directory))
                     break
